requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Environment Variables
python-dotenv==1.0.1

# Utilities
pydantic==2.6.1
python-dateutil==2.8.2
